requirements.txt
boto3==1.24.42
debugpy==1.6.2
dnspython==2.2.1
filelock==3.7.1
frozendict==2.3.4
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
Jinja2===3.1.2
junit-report==0.2.7
kubernetes==24.2.0
libvirt-python==8.6.0
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
ntnx-api>=1.5.0
openshift-client==1.0.17
paramiko==2.7.2
pre-commit==2.20.0
pycharm_remote_debugger==0.1.13
pytest-xdist==2.5.0
pytest==7.1.2
python-dateutil==2.8.2
python-hcl2==3.0.5
python-terraform==0.10.1
pyyaml==6.0
requests==2.25.1
retry==0.9.2
scp==0.14.4
semver==2.13.0
tabulate==0.8.10
tqdm==4.64.0
urllib3==1.26.4
pyvmomi>=7.0.2
waiting>=1.4.1
prompt_toolkit==3.0.30
